requirements.txt
# MaxPark RFID Access Control System - Python Dependencies
# Core web framework
Flask==2.3.3
Werkzeug==2.3.7

# Raspberry Pi GPIO control
RPi.GPIO==0.7.1
pigpio==1.78

# Computer vision and image processing
opencv-python==4.8.1.78

# Firebase integration
firebase-admin==6.2.0
google-api-core==2.11.1
google-cloud-firestore==2.11.1

# HTTP requests and API communication
requests==2.31.0
urllib3==2.0.7

# Environment variable management
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Type hints support
typing-extensions==4.8.0

# Additional security and cryptography
cryptography==41.0.7

# Image processing and manipulation
Pillow==10.0.1

# System monitoring and process management
psutil==5.9.6

# I2C communication for DS3232 RTC
smbus2==0.4.3

# JSON schema validation (optional, for API validation)
jsonschema==4.19.2

# CORS support for web interface (optional)
Flask-CORS==4.0.0

# Development and testing dependencies (optional)
# pytest==7.4.3
# pytest-flask==1.2.0
# black==23.9.1
# flake8==6.1.0

# Production server (optional - for production deployment)
# gunicorn==21.2.0
# waitress==2.1.2

# System service management (optional)
# systemd-python==235

# Network utilities (optional)
# netifaces==0.11.0

# File system monitoring (optional)
# watchdog==3.0.0

# Database support (optional - if extending to use databases)
# SQLAlchemy==2.0.23
# Flask-SQLAlchemy==3.1.1

# Caching support (optional)
# Flask-Caching==2.1.0
# redis==5.0.1

# Background task processing (optional)
# celery==5.3.4
# redis==5.0.1

# Logging enhancements (optional)
# structlog==23.2.0

# Configuration management (optional)
# dynaconf==3.2.4

# API documentation (optional)
# flask-restx==1.3.0
# apispec==6.3.0

# WebSocket support (optional)
# flask-socketio==5.3.6

# Real-time communication (optional)
# socketio==5.8.0

# Data validation (optional)
# marshmallow==3.20.1
# flask-marshmallow==0.15.0

# File upload handling (optional)
# flask-uploads==0.2.1

# Email notifications (optional)
# flask-mail==0.9.1

# Backup and restore (optional)
# boto3==1.29.7
# paramiko==3.3.1

# Performance monitoring (optional)
# flask-profiler==1.8.1

# Health checks (optional)
# flask-healthz==1.0.1

# Rate limiting (optional)
# flask-limiter==3.5.0

# Security headers (optional)
# flask-talisman==1.1.0

# Compression (optional)
# flask-compress==1.13

# Static file serving (optional)
# flask-static-compress==1.0.0

# Template caching (optional)
# flask-caching==2.1.0

# Internationalization (optional)
# flask-babel==4.0.0

# Admin interface (optional)
# flask-admin==1.6.1

# User management (optional)
# flask-user==1.0.2.2

# OAuth integration (optional)
# flask-oauthlib==0.9.6

# JWT authentication (optional)
# PyJWT==2.8.0
# flask-jwt-extended==4.5.3

# QR code generation (optional)
# qrcode==7.4.2

# Barcode generation (optional)
# python-barcode==0.15.1

# PDF generation (optional)
# reportlab==4.0.7

# Excel file handling (optional)
# openpyxl==3.1.2
# xlsxwriter==3.1.9
